Move `if_terminate` to pure optimizer
